requirements.txt
# $ pip3 install -r requirements.txt

# Production dependencies
pandas
python-dotenv
langchain_openai

# Development dependencies
pytest